+++ packages/web/src/views/governance/components/governance-summary/governance-detail/GovernanceDetail.styles.ts
