+++ packages/web/src/views/governance/components/governance-summary/governance-detail/governance-detail-info/GovernanceDetailInfo.styles.ts
