+++ packages/web/src/views/governance/components/proposals-list/proposal-detail/ProposalDetail.spec.tsx
