+++ packages/web/src/views/governance/components/proposals-list/proposal-detail/ProposalDetailSekeleton.tsx
