+++ packages/web/src/views/governance/components/proposals-list/proposal-header/ProposalHeader.styles.ts
