+++ packages/web/src/views/governance/components/proposals-list/view-proposal-modal/ViewProposalModal.stories.tsx
